requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.5; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
et-xmlfile==2.0.0; python_version >= '3.8'
iniconfig==2.0.0; python_version >= '3.7'
numpy==2.2.2; python_version >= '3.10'
openpyxl==3.1.5; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pluggy==1.5.0; python_version >= '3.8'
psycopg2==2.9.10; python_version >= '3.8'
pytest==8.3.4; python_version >= '3.8'
pytest-django==4.9.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.1
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.3; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
